# docs/contracts/qalem-document-provider-v1.openapi.yaml
openapi: 3.1.0
info:
  title: Diwan, fournisseur documentaire pour Qalem
  version: "1.0"
  description: >
    Facade documentaire versionnee. Qalem consomme cette API et rien d'autre :
    aucun acces direct a SurrealDB, aucun vecteur transmis, aucune recherche
    globale. L'organisation autorisee est toujours derivee du jeton presente.
servers:
  - url: https://diwan.ai-mpower.com
security:
  - serviceToken: []
paths:
...
  /api/v1/consumers/qalem/connectors/google-drive/callback:
    get:
      summary: Terminer l'autorisation Google Drive
      security: []
      parameters:
        - { in: query, name: code, required: true, schema: { type: string } }
        - { in: query, name: state, required: true, schema: { type: string } }
      responses:
        "303": { description: Redirection vers Qalem }
        "400": { $ref: "#/components/responses/Erreur" }
        "503": { $ref: "#/components/responses/Erreur" }
...
components:
  securitySchemes:
    serviceToken:
      type: http
      scheme: bearer
      description: >
        Jeton de service propre a Qalem, distinct du mot de passe global
        historique de Diwan. Diwan n'en conserve que l'empreinte SHA-256.
  responses:
    Erreur:
      description: Erreur au format stable du contrat
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      required: [contractVersion, requestId, error]
      properties:
        contractVersion: { type: string }
        requestId: { type: string }
        error:
          type: object
          required: [code, message, retryable, details]
          properties:
            code:
              type: string
              enum:
                - UNAUTHENTICATED
                - FORBIDDEN
                - CORPUS_NOT_FOUND
                - SOURCE_NOT_FOUND
                - SOURCE_NOT_AUTHORIZED
                - SOURCE_NOT_READY
                - INVALID_SOURCE_TYPE
                - FILE_TOO_LARGE
                - INGESTION_FAILED
                - EXTRACTION_FAILED
                - OCR_FAILED
                - EMBEDDING_FAILED
                - EMBEDDING_MODEL_UNAVAILABLE
                - EMBEDDING_DIMENSION_MISMATCH
                - INSUFFICIENT_EVIDENCE
                - RATE_LIMITED
                - SERVICE_BUSY
                - CONTRACT_VERSION_UNSUPPORTED
                - INVALID_REQUEST
                - CONNECTION_REQUIRED
                - OAUTH_STATE_INVALID
                - EXTERNAL_SOURCE_NOT_FOUND
                - PROVIDER_UNAVAILABLE
            message: { type: string }
            retryable: { type: boolean }
            details: { type: object }
...